chips added, features state moved to common interface

+++ composeApp/src/commonMain/kotlin/features/tasks/completed/CompletedTasksContainer.kt
@@ -1,6 +1,6 @@
 package features.tasks.completed
 
-import features.tasks.TasksState
+import features.OTrackerState
 import flow_mvi.DefaultConfigurationFactory
 import flow_mvi.configure
 import pro.respawn.flowmvi.api.Container
@@ -10,50 +10,60 @@ import pro.respawn.flowmvi.plugins.recover
 import pro.respawn.flowmvi.plugins.reduce
 import ru.kpfu.itis.common.mapper.ErrorMapper
 import ru.kpfu.itis.features.task.data.repository.TaskRepository
+import ru.kpfu.itis.features.task.data.store.UserStore
 import ru.kpfu.itis.features.task.domain.model.TaskModel
 
 class CompletedTasksContainer(
     private val errorMapper: ErrorMapper,
     private val repository: TaskRepository,
     private val configurationFactory: DefaultConfigurationFactory,
-) : Container<TasksState<List<TaskModel>>, CompletedTasksIntent, CompletedTasksAction> {
+    private val userStore: UserStore,
+) : Container<OTrackerState<List<TaskModel>>, CompletedTasksIntent, CompletedTasksAction> {
 
-    override val store: Store<TasksState<List<TaskModel>>, CompletedTasksIntent, CompletedTasksAction> =
-        store(TasksState.Initial) {
+    override val store: Store<OTrackerState<List<TaskModel>>, CompletedTasksIntent, CompletedTasksAction> =
+        store(OTrackerState.Initial) {
+
 
             configure(configurationFactory, "Tasks")
 
             recover { exception ->
                 updateState {
-                    TasksState.Error(errorMapper.map(exception = exception))
+                    OTrackerState.Error(errorMapper.map(exception = exception))
                 }
                 null
             }
 
             reduce { intent ->
-                when (intent) {
-                    is CompletedTasksIntent.LoadTasks -> {
-                        updateState { TasksState.Loading }
-                        val tasks = loadActiveTasks(2)
-                        updateState { TasksState.Success(tasks) }
-                    }
 
-                    is CompletedTasksIntent.EditTask -> {
-                        val task = repository.getTask(intent.taskId)
-                        action(CompletedTasksAction.OpenTaskBottomSheet(task.id))
-                    }
+                val userId = userStore.getUserId()
+                if (userId == null) {
+                    action(CompletedTasksAction.Logout)
+                    intent(CompletedTasksIntent.ClearUserCache)
+                } else {
+                    when (intent) {
+                        is CompletedTasksIntent.LoadTasks -> {
+                            updateState { OTrackerState.Loading }
+                            val tasks = repository.getActiveTasks(userId)
+                            updateState { OTrackerState.Success(tasks) }
+                        }
 
-                    is CompletedTasksIntent.DeleteTask -> {
-                        updateState { TasksState.Loading }
-                        repository.deleteTask(intent.taskId)
-                        val tasks = loadActiveTasks(2)
-                        updateState { TasksState.Success(tasks) }
+                        is CompletedTasksIntent.EditTask -> {
+                            val task = repository.getTask(intent.taskId)
+                            action(CompletedTasksAction.OpenTaskBottomSheet(task.id))
+                        }
+
+                        is CompletedTasksIntent.DeleteTask -> {
+                            updateState { OTrackerState.Loading }
+                            repository.deleteTask(intent.taskId)
+                            val tasks = repository.getActiveTasks(userId)
+                            updateState { OTrackerState.Success(tasks) }
+                        }
+
+                        is CompletedTasksIntent.ClearUserCache -> {
+                            userStore.deleteUserId()
+                        }
                     }
                 }
             }
         }
-
-    private suspend fun loadActiveTasks(userId: Long): List<TaskModel> {
-        return repository.getActiveTasks(userId)
-    }
 }
